% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_connect_to_db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{connect_to_db}
\usage{
connect_to_db()
}
\value{
un objet de type connexion à la base
}
\description{
une fonction pour se connecter à la bdd en mariadb
}
\examples{
con <- connect_to_db()
}
